We now have an implementation of strptime for Windows. Two minor changes to the original strptime code so that it compiles for us.

+++ common/utils/strptime.c
@@ -32,7 +32,7 @@
 
 #ifdef WIN32
 
-#include "platform/platform.h"
+/* #include "platform/platform.h" */
 
 
 static const char *abb_weekdays[] = {
@@ -106,7 +106,7 @@ match_string (const char **buf, const char **strs)
     int i = 0;
 
     for (i = 0; strs[i] != NULL; ++i) {
-	int len = strlen (strs[i]);
+	size_t len = strlen (strs[i]);
 
 	if (strncasecmp (*buf, strs[i], len) == 0) {
 	    *buf += len;
